pyproject.toml
[project]
name = "website-comparison-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "flask>=3.1.0",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "requests>=2.32.3",
    "scikit-image>=0.25.2",
    "selenium>=4.32.0",
    "webdriver-manager>=4.0.2",
]
